+++ CHANGELOG.md
@@ -95,6 +95,7 @@ This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 - (RE-14305) Add 'vanagon dependencies' command to generate gem dependencies as a json file
 - (VANAGON-162) Added new instance variable 'log_url' to use in logs rather than the full git url
 - (maint) Check environment for the X-RPROXY-PASS variable and add it to the http request header in the download method if it exists
+- (maint) Allow target on CLI to set user and port
 
 ### Removed
 - (VANAGON-168) Remove Fedora 30 x86_64